requirements.txt
beautifulsoup4
requests
langchain
llama-index
gradio
sentence-transformers
numpy
faiss-cpu
